Rename to opentelemetry-exporter-sender-grpc-managed-channel
