+++ docs/apidiffs/current_vs_latest/opentelemetry-exporter-sender-grpc-managed-channel.txt
