+++ exporters/sender/grpc-managed-channel/src/main/resources/META-INF/services/io.opentelemetry.exporter.internal.grpc.GrpcSenderProvider
@@ -1,0 +1,1 @@
+io.opentelemetry.exporter.sender.grpc.managedchannel.internal.UpstreamGrpcSenderProvider
