--- exporters/sender/grpc-upstream/src/main/resources/META-INF/services/io.opentelemetry.exporter.internal.grpc.GrpcSenderProvider
@@ -1,1 +1,0 @@
-io.opentelemetry.exporter.sender.grpc.upstream.internal.UpstreamGrpcSenderProvider
