+++ settings.gradle.kts
@@ -32,7 +32,7 @@ include(":extensions:incubator")
 include(":extensions:kotlin")
 include(":extensions:trace-propagators")
 include(":exporters:common")
-include(":exporters:sender:grpc-upstream")
+include(":exporters:sender:grpc-managed-channel")
 include(":exporters:sender:jdk")
 include(":exporters:sender:okhttp")
 include(":exporters:jaeger")
